refactor[db]: move db file
